+++ evalml/tests/data_checks_tests/test_data_checks.py
@@ -18,6 +18,7 @@ from evalml.data_checks import (
     DateTimeFormatDataCheck,
     DefaultDataChecks,
     TargetDistributionDataCheck,
+    TimeSeriesParametersDataCheck,
 )
 from evalml.exceptions import DataCheckInitError
 from evalml.problem_types import ProblemTypes, is_time_series
@@ -500,7 +501,10 @@ def test_default_data_checks_across_problem_types(problem_type):
     default_data_check_list = DefaultDataChecks._DEFAULT_DATA_CHECK_CLASSES
 
     if is_time_series(problem_type):
-        default_data_check_list = default_data_check_list + [DateTimeFormatDataCheck]
+        default_data_check_list = default_data_check_list + [
+            DateTimeFormatDataCheck,
+            TimeSeriesParametersDataCheck,
+        ]
     if problem_type in [
         ProblemTypes.REGRESSION,
         ProblemTypes.TIME_SERIES_REGRESSION,
@@ -511,14 +515,35 @@ def test_default_data_checks_across_problem_types(problem_type):
     else:
         default_data_check_list = default_data_check_list + [ClassImbalanceDataCheck]
 
+    problem_config = {
+        "gap": 1,
+        "max_delay": 1,
+        "forecast_horizon": 1,
+        "date_index": "datetime",
+    }
     data_check_classes = [
         check.__class__
         for check in DefaultDataChecks(
-            problem_type, get_default_primary_search_objective(problem_type)
+            problem_type,
+            get_default_primary_search_objective(problem_type),
+            problem_configuration=problem_config
+            if is_time_series(problem_type)
+            else None,
         ).data_checks
     ]
 
     assert data_check_classes == default_data_check_list
+
+
+def test_default_data_checks_missing_problem_configuration_for_time_series(ts_data):
+    with pytest.raises(
+        ValueError,
+        match="problem_configuration cannot be None for time series problems!",
+    ):
+        DefaultDataChecks(
+            "time series binary",
+            get_default_primary_search_objective("time series regression"),
+        )
 
 
 def test_data_checks_init_from_classes():
@@ -651,8 +676,16 @@ def test_errors_warnings_in_invalid_target_data_check(objective, ts_data):
         details=details,
     ).to_dict()
 
+    problem_config = {
+        "gap": 1,
+        "max_delay": 1,
+        "forecast_horizon": 1,
+        "date_index": "datetime",
+    }
     default_data_check = DefaultDataChecks(
-        problem_type="time series regression", objective=objective
+        problem_type="time series regression",
+        objective=objective,
+        problem_configuration=problem_config,
     ).data_checks
     for check in default_data_check:
         if check.name == "InvalidTargetDataCheck":