+++ docs/source/release_notes.rst
@@ -3,6 +3,7 @@ Release Notes
 **Future Releases**
     * Enhancements
     * Fixes
+        * Fixed error caused when tuning threshold for time series binary classification :pr: ``
     * Changes
     * Documentation Changes
     * Testing Changes